refactor: kill struct and reorganise src :file_cabinet:
