+++ examples/async-with-shortstop.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const Configue = require('../lib/configue');
 
 const configueOptions = {
