+++ examples/basic-with-files.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const Configue = require('configue');
 
 const configueOptions = {
