+++ examples/hapi-server.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const Hapi = require('hapi');
 const Configue = require('configue');
