+++ package.json
@@ -16,7 +16,7 @@
   "engines": {
     "node": ">=8.10.0"
   },
-  "main": "src/configue.js",
+  "main": "src/index.js",
   "author": "Adrien Becchis (https://github.com/AdrieanKhisbe)",
   "maintainers": [
     "Adrien Becchis (https://github.com/AdrieanKhisbe)"
