+++ src/core/builder.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const _ = require('lodash');
 
 const optionKeys = [
