+++ src/core/utils.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const _ = require('lodash/fp');
 
 const getPaths = (obj, basePath = []) =>
